pyproject.toml
[tool.poetry]
name = "mergics"
version = "0.1.0"
description = "Merge multiple ICS streams"
authors = ["Guilhem Saurel <guilhem.saurel@laas.fr>"]
license = "BSD-2-Clause"

[tool.poetry.dependencies]
python = "^3.8"
ndh = "^3.10.4"
icalendar = "^4.0.7"
psycopg2 = "^2.8.6"
python-memcached = "^1.59"
requests = "^2.24.0"
gunicorn = "^20.0.4"
django-yeouia = "^2.0.0"

[tool.poetry.dev-dependencies]
ipython = "^7.18.1"
flake8 = "^3.8.4"
black = "^22.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
